Add cast to suppress compiler complaint.

+++ src/block/block_open.c
@@ -185,7 +185,8 @@ __wt_block_open(WT_SESSION_IMPL *session, const char *filename, WT_BLOCK_FILE_OP
     block->allocfirst = WT_STRING_MATCH("first", cval.str, cval.len);
     block->has_objects = (opener != NULL);
     if (block->has_objects)
-        block->logid = opener->current_object_id(opener);
+        /* TODO: tiered: fix 32 bit vs 64 bit mismatch. */
+        block->logid = (uint32_t)opener->current_object_id(opener);
 
     /* Configuration: optional OS buffer cache maximum size. */
     WT_ERR(__wt_config_gets(session, cfg, "os_cache_max", &cval));
